+++ CHANGES.md
@@ -11,10 +11,12 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
-CHANGES  4.2.3   ??/08/2024
+CHANGES  4.2.3   ??/09/2024
 ---------------------------
     * add \_rl\_repr \_rl\_docdent functions
     * fix test outputs to be 3.13 compatible 
+    * Use lru_cache in axes.py
+    * fix keyword arguments in renderPDF.py and usage in shapes.py
 
 CHANGES  4.2.2   21/06/202
 ---------------------------